Adding `posterior_transform` to `qMaxValueEntropySearch`'s input constructor (#2181)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/2181

This diff adds `posterior_transform` to the input constructor of the `qMaxValueEntropySearch` acquisition function. This can be used to transform the posterior distribution before computing the maximum entropy search, a property that Ax relies on to encode whether an objective is to be maximized or minimized, instead of passing the `minimize` keyword, which is not accepted by  all acquisition functions.

This solves https://github.com/facebook/Ax/issues/2133.

Reviewed By: esantorella

Differential Revision: D53101380

fbshipit-source-id: 80aa8581663e295d05610b6751f0df7b3e300e20